toying with keeping --raw-css in sync
it appears doing this at compile-time is a bad idea, it should definitely be a plugin, but there's a fair amount to do here because just applying the raw css won't work with em & var values most the time.

This is I think a really killer feature that you'll see all instances of that value & preview them, and click them to be taken to the places in the dom, or css where those things are applied.

Hmm.. very interesting & definitely going to be one of the great features